services/notification-python/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic[email]==2.9.2
aio_pika==9.4.3
httpx==0.27.2
python-dotenv==1.0.1

